Rectangle Tool
Created a cShape parent-class and a cRectangle child-class which is the rectangle tool. Successfully created a cRectangle object and tested. The rectangle tool functionality remained the same! Able to resize and draw. However, this is only specific to the rectangle tool class. Looking to create a dynamic and smart Toolbox class that allows the user to select their tool.