Initial OOP Structure
Class Structure:
 - cWindow (Extends from sf::RenderWindow)
 - cCanvas
 - cEventManager
 - cToolManager
 - cTool (Parent)
    > cRectangle (Child)

Changes:
Basically moved everything into their own classes that dynamically handles different things for simple coding in the future. First the cWindow class extends from the sf::RenderWindow class from the SFML library. It calls and instantiates the cEventManager class which dynamically handles events as well as realtime input. Events call for the cToolManager class which allows for dynamic tool selection and usage. The cTool parent class is essentially a tool blueprint for any future tools created. cRectangle child class is an example of using the cTool class as a blueprint.